Fixed recursive infinite looping in`toString()` of  `Database`. Improved `DataUpdate`, now containing a list of `DBEntry`s instead of only the compiled bytes. Be more verbose with output.
# Recursive infinite looping:
Happened in `toString()` of database containing update Listeners containing information about the same database, leading to another time the listeners, and thus the database, and so on...
Fixed by Introducing a `toStringWithoutDB()` into the interface `IDataUpdateListener`, and using that instead when printing the database UpdateListeners.

# Improved `DataUpdate`:
Now has a list of `DBEntry`s instead of only the compiled bytes. This allows you to see the content of the package as verbose String.
The bytes gets dynamically generated inside `getBytes()`.

# Be more verbose with output:
The `Server` now prints when threads like Heartbeat die.